Fix parsing of initial states of non-deterministic automaton.
The application was reading the initial states as a single string instead of
a set of states. The code that created the initial configurations iterated over
the initial states variable, and, as it was a string (instead of a list of
states/strings), it was considering each letter as a single state.